pyproject.toml
[tool.poetry]
name = "arcanumbot"
version = "2.1.0"
description = "Discord bot for the arcanum"
authors = ["StarrFox <starrfox6312@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
jishaku = "^2.5.1"
pendulum = "^2.1.2"
numpy = "^1.24.3"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}
humanize = "^4.6.0"
"discord.py" = "^2.3.0"
psutil = "^5.9.5"
click = "^8.1.3"
uvloop = {version = "^0.17.0", markers = "platform_system == 'Linux'"}
asyncpg = "^0.27.0"

[tool.poetry.scripts]
arcanumbot = "arcanumbot.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
